1. 예외 생성자 추가 2. 서비스 레이어에 회원 가입 로직 추가. 예외 발생시 던짐. 3. 컨트롤러와 레스트 컨트롤러 분리.
