+++ src/main/java/com/dasd412/remake/api/controller/exception/DuplicateEmailException.java
@@ -1,7 +1,12 @@
 package com.dasd412.remake.api.controller.exception;
 
 //OAuth 가 아닌 회원 가입 시 이메일이 중복되면 발생합니다.
-public class DuplicateEmailException extends DuplicateException{
+public class DuplicateEmailException extends DuplicateException {
+
+    public DuplicateEmailException(String message) {
+        super(message);
+    }
+
     public DuplicateEmailException(String message, Throwable cause) {
         super(message, cause);
     }
