+++ src/main/java/com/dasd412/remake/api/service/security/WriterService.java
@@ -1,5 +1,7 @@
 package com.dasd412.remake.api.service.security;
 
+import com.dasd412.remake.api.controller.exception.DuplicateEmailException;
+import com.dasd412.remake.api.controller.exception.DuplicateUserNameException;
 import com.dasd412.remake.api.domain.diary.EntityId;
 import com.dasd412.remake.api.domain.diary.writer.Role;
 import com.dasd412.remake.api.domain.diary.writer.Writer;
@@ -34,10 +36,27 @@ public class WriterService {
         return EntityId.of(Writer.class, writerId + 1);
     }
 
+    //비밀 번호 암호화
+    private String encodePassword(String rawPassword) {
+        return bCryptPasswordEncoder.encode(rawPassword);
+    }
+
     @Transactional
     public void saveWriterWithSecurity(String name, String email, String password, Role role) {
-        logger.info("saveWriter");
-        String encodedPassword = bCryptPasswordEncoder.encode(password);
+        logger.info("join writer with security");
+
+        if (writerRepository.existsName(name) == Boolean.TRUE) {
+            throw new DuplicateUserNameException("이미 존재하는 회원 이름입니다.");
+        }
+
+        if (writerRepository.existsEmail(email) == Boolean.TRUE) {
+            throw new DuplicateEmailException("이미 존재하는 이메일입니다.");
+        }
+
+        //비밀 번호 암호화
+        String encodedPassword = encodePassword(password);
+
+        //실제 회원 가입
         Writer writer = new Writer(getNextIdOfWriter(), name, email, encodedPassword, role);
         writerRepository.save(writer);
     }
